Refactored console command tests to use Application tester.
